feat: split `noteit.board.es.js` into files with each classes
